+++ conan/api/subapi/config.py
@@ -34,7 +34,7 @@ class ConfigAPI:
     def install(self, path_or_url, verify_ssl, config_type=None, args=None,
                 source_folder=None, target_folder=None):
         # TODO: We probably want to split this into git-folder-http cases?
-        from conans.client.conf.config_installer import configuration_install
+        from conan.internal.api.config.config_installer import configuration_install
         app = ConanApp(self.conan_api)
         configuration_install(app, path_or_url, verify_ssl, config_type=config_type, args=args,
                               source_folder=source_folder, target_folder=target_folder)
@@ -89,7 +89,7 @@ class ConfigAPI:
                                        "skipping configuration install")
                     return pkg.pref  # Already installed, we can skip repeating the install
 
-        from conans.client.conf.config_installer import configuration_install
+        from conan.internal.api.config.config_installer import configuration_install
         configuration_install(app, uri=pkg.conanfile.package_folder, verify_ssl=False,
                               config_type="dir", ignore=["conaninfo.txt", "conanmanifest.txt"])
         # We save the current package full reference in the file for future
