+++ conans/client/rest_client_local_recipe_index.py
@@ -9,7 +9,7 @@ import yaml
 from conan.api.model import LOCAL_RECIPES_INDEX
 from conan.api.output import ConanOutput
 from conan.internal.cache.home_paths import HomePaths
-from conans.client.cmd.export import cmd_export
+from conan.internal.api.export import cmd_export
 from conans.client.loader import ConanFileLoader
 from conans.errors import ConanException, PackageNotFoundException, RecipeNotFoundException, \
     ConanReferenceDoesNotExistInDB, NotFoundException
