+++ test/functional/command/config_install_test.py
@@ -8,7 +8,7 @@ import pytest
 from mock import patch
 
 from conan.api.model import Remote
-from conans.client.conf.config_installer import _hide_password
+from conan.internal.api.config.config_installer import _hide_password
 from conans.client.downloaders.file_downloader import FileDownloader
 from conan.internal.paths import DEFAULT_CONAN_HOME
 from conan.test.assets.genconanfile import GenConanfile
