+++ test/functional/command/profile_test.py
@@ -5,7 +5,7 @@ import unittest
 
 import pytest
 
-from conans.client.conf.detect import detect_defaults_settings
+from conan.internal.api.profile.detect import detect_defaults_settings
 from conan.test.utils.mocks import RedirectedTestOutput
 from conan.test.utils.tools import TestClient, redirect_output
 from conan.test.utils.env import environment_update
